entegration has started

+++ PublicTransportApp/Models/Stops/Transfer.cs
@@ -4,6 +4,6 @@
     {
         public string TransferStopId { get; set; }
         public int TransferSure { get; set; }
-        public double TransferUcret { get; set; }
+        public float TransferUcret { get; set; }   // it was double    
     }
 }
